Added UserRole n CRUD functionality
