+++ Backend/LearningAPI/LearningAPI/MyDbContext.cs
@@ -18,5 +18,6 @@ namespace LearningAPI
         public required DbSet<Tutorial> Tutorials { get; set; }
 
         public required DbSet<User> Users { get; set; }
+        public required DbSet<UserRole> UserRoles { get; set; }
     }
 }
